+++ crates/px/src/commands/project/stack.rs
@@ -1,7 +1,14 @@
+use std::path::{Path, PathBuf};
+
 use anyhow::Result;
 use px_cli::StackProjectArgs;
+use siril_sys::{
+    BestRejection, Builder, FitsExt,
+    commands::{Convert, Load, Register, SeqApplyReg, Stack},
+    siril_ext::{CdExt, MirrorxExt, SaveExt},
+};
 
-use crate::{ExitStatus, printer::Printer};
+use crate::{ExitStatus, printer::Printer, utils::wait_for_confirm};
 
 pub(crate) async fn stack_project_observations(
     args: StackProjectArgs,
@@ -22,5 +29,110 @@ pub(crate) async fn stack_project_observations(
         config
     ))?;
 
+    for stack in config.linear_stacks {
+        // Setup siril
+        let ext = FitsExt::FIT;
+        let mut siril = Builder::default()
+            .output_sink(siril_sys::OutputSink::Inherit)
+            .use_extension(ext.clone())
+            .build()
+            .await?;
+
+        // manage the sequence
+        let mut prefix = String::from("light_");
+
+        // convert each input directory
+        let mut start_idx = 1;
+        for obs in stack.observations {
+            let obs_path = raw_input_to_output_folder(&obs.path);
+            let count = std::fs::read_dir(&obs_path)?
+                .filter_map(|e| e.ok())
+                .filter(|e| {
+                    e.path()
+                        .extension()
+                        .is_some_and(|x| x.eq_ignore_ascii_case(ext.to_string().as_str()))
+                })
+                .count();
+            siril.cd(obs_path).await?;
+            siril
+                .execute(
+                    &Convert::builder(&prefix)
+                        .output_dir(siril.initial_directory())
+                        .start_index(start_idx)
+                        .build(),
+                )
+                .await?;
+            start_idx += count as u8;
+        }
+
+        // Return to working directory
+        siril.cd(siril.initial_directory()).await?;
+
+        // TODO: Optional: run bg extraction on every frame before stacking
+        // if extract_background:
+        //     await siril.command(seqsubsky(prefix))
+        //     prefix = f"bkg_{prefix}"
+
+        // Register all the images
+        siril
+            .execute(&Register::builder(&prefix).two_pass(true).build())
+            .await?;
+
+        // Generate their transformed version
+        siril
+            .execute(&SeqApplyReg::builder(&prefix).build())
+            .await?;
+        prefix = format!("r_{prefix}");
+
+        // Find the best rejection method
+        let rejection = BestRejection::find(start_idx as usize);
+        printer.info(format!("Found best stacking rejection: {:?}", rejection))?;
+
+        // Stack the background extracted images
+        siril
+            .execute(
+                &Stack::builder(prefix)
+                    .norm(siril_sys::StackNormFlag::AddScale)
+                    .filter_included(true)
+                    .output_norm(true)
+                    .rgb_equalization(true)
+                    .rejection(rejection.method)
+                    .lower_rej(rejection.low_threshold)
+                    .higher_rej(rejection.high_threshold)
+                    .out("result")
+                    .build(),
+            )
+            .await?;
+
+        // Load and flip the image if needed
+        siril.execute(&Load::builder("result").build()).await?;
+        siril.mirrorx(true).await?;
+        siril.execute(&Load::builder("result").build()).await?;
+
+        // TODO: output file naming and config for HDR
+        // TODO: include date?
+        // TODO: Save this output file name to the project config?
+
+        // Output file for the linear_stack
+        let filter_output_file = project_dir.join(format!("{}_linear_stack", stack.filter));
+        siril.save(filter_output_file).await?;
+
+        // TODO: Split and save RGB from OSC image
+
+        printer.success(format!("{} linear stack complete", stack.filter))?;
+
+        wait_for_confirm(printer).await;
+    }
+
     Ok(ExitStatus::Success)
+}
+
+fn raw_input_to_output_folder(path: &Path) -> PathBuf {
+    let new_name = path
+        .file_name()
+        .unwrap()
+        .to_str()
+        .unwrap()
+        .replacen("RAW_", "PP_", 1);
+    path.with_file_name(new_name)
 }